// openapi/cineast/model/mediaSegmentQueryResult.ts
/**
 * Cineast RESTful API
 * Cineast is vitrivr\'s content-based multimedia retrieval engine. This is it\'s RESTful API.
 *
 * The version of the OpenAPI document: v1
 * Contact: contact@vitrivr.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MediaSegmentDescriptor } from './mediaSegmentDescriptor';


export interface MediaSegmentQueryResult { 
    content?: Array<MediaSegmentDescriptor>;
    queryId?: string;
    messageType?: MediaSegmentQueryResult.MessageTypeEnum;
}
export namespace MediaSegmentQueryResult {
    export type MessageTypeEnum = 'PING' | 'Q_SIM' | 'Q_MLT' | 'Q_NESEG' | 'Q_SEG' | 'M_LOOKUP' | 'Q_TEMPORAL' | 'SESSION_START' | 'QR_START' | 'QR_END' | 'QR_ERROR' | 'QR_OBJECT' | 'QR_METADATA_O' | 'QR_METADATA_S' | 'QR_SEGMENT' | 'QR_SIMILARITY';
    export const MessageTypeEnum = {
        PING: 'PING' as MessageTypeEnum,
        QSIM: 'Q_SIM' as MessageTypeEnum,
        QMLT: 'Q_MLT' as MessageTypeEnum,
        QNESEG: 'Q_NESEG' as MessageTypeEnum,
        QSEG: 'Q_SEG' as MessageTypeEnum,
        MLOOKUP: 'M_LOOKUP' as MessageTypeEnum,
        QTEMPORAL: 'Q_TEMPORAL' as MessageTypeEnum,
        SESSIONSTART: 'SESSION_START' as MessageTypeEnum,
        QRSTART: 'QR_START' as MessageTypeEnum,
        QREND: 'QR_END' as MessageTypeEnum,
        QRERROR: 'QR_ERROR' as MessageTypeEnum,
        QROBJECT: 'QR_OBJECT' as MessageTypeEnum,
        QRMETADATAO: 'QR_METADATA_O' as MessageTypeEnum,
        QRMETADATAS: 'QR_METADATA_S' as MessageTypeEnum,
        QRSEGMENT: 'QR_SEGMENT' as MessageTypeEnum,
        QRSIMILARITY: 'QR_SIMILARITY' as MessageTypeEnum
    };
}